Load departments

+++ complete/src/main/java/com/example/accessingdatarest/repository/RegionRepository.java
@@ -1,11 +1,13 @@
 package com.example.accessingdatarest.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import com.example.accessingdatarest.entity.Region;
 
 @RepositoryRestResource(collectionResourceRel = "regions", path = "regions")
-public interface RegionRepository extends PagingAndSortingRepository<Region, Long> {
+public interface RegionRepository extends PagingAndSortingRepository<Region, BigInteger> {
 
 }
